+++ src/routes/login/[magicLink]/+page.server.ts
@@ -5,12 +5,7 @@ export const load = async ({ params, cookies }) => {
 	// Set the magic link cookie with a 90 day expiry
 	// This should cover the longest of registration periods
 	cookies.set('magicLink', params.magicLink, { path: '/', maxAge: 60 * 60 * 24 * 90 });
-	// await trpc(cookies).users.verify();
-	// throw redirect(303, '/');
-
-	try {
-		await trpc(cookies).users.verify();
-	} catch (error) {
+	if ((await trpc(cookies).users.get()) === null) {
 		throw redirect(303, '/?magiclink');
 	}
 	throw redirect(303, '/');